Create find sub array with least average file.
References Issue #16. 
The algorithm finds the subarray with least average. It runs through the array until arr.length - k, calculates the average given the elements between the index and k + index. If the average calculated is smaller than the min defined, it is replaced. The minimum sub array is returned